requirements.txt
# Fusion Prime - Root Requirements
# This file consolidates all service dependencies for CI compatibility

# Core dependencies
fastapi==0.112.0
uvicorn>=0.38.0
pydantic==2.12.3
httpx==0.28.1
requests==2.32.4

# Database
sqlalchemy>=2.0.44
psycopg2-binary>=2.9.0
aiosqlite>=0.21.0
asyncpg>=0.30.0

# Google Cloud
google-cloud-pubsub>=2.31.0
google-cloud-storage>=2.10.0
google-auth>=2.41.0

# Web3 and blockchain
web3>=7.14.0
eth-account>=0.13.0
eth-utils>=5.3.0

# Testing
pytest==8.4.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-html==4.1.0
pytest-xdist==3.5.0

# Code quality
black>=24.10.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.18.0
ruff>=0.5.0

# Security
bandit>=1.8.0
safety>=3.6.0

# Utilities
python-dotenv>=1.1.0
pyyaml>=6.0.0
click>=8.3.0
tenacity>=9.1.0
